# heka-identity-service/docs/swagger-spec.yaml
openapi: 3.0.0
paths:
...
  /credential-definitions:
    get:
      operationId: CredentialDefinitionController_find
      parameters:
        - name: issuerId
          required: true
          in: query
          schema:
            type: string
        - name: schemaId
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Credential Definition Records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CredentialDefinitionDto"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      security: &a22
        - bearer: []
      summary: Get created credential definitions
      tags: &a23
        - Credential Definitions
    post:
      operationId: CredentialDefinitionController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCredentialDefinitionDto"
      responses:
        "201":
          description: Credential Definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialDefinitionDto"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "422":
          description: Unprocessable Entity
      security: *a22
      summary: Create a credential definition
      tags: *a23
  /credential-definitions/{id}:
    get:
      operationId: CredentialDefinitionController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Credential Definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialDefinitionDto"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security: *a22
      summary: Get a credential definition
      tags: *a23
  /connections:
    get:
      operationId: ConnectionController_find
      parameters: []
      responses:
        "200":
          description: Connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConnectionRecordDto"
        "401":
          description: Unauthorized
      security: &a24
        - bearer: []
      summary: Get all connections
      tags: &a25
        - Connections
  /connections/create-invitation:
    post:
      operationId: ConnectionController_createInvitation
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationRequestDto"
      responses:
        "200":
          description: Invitation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateInvitationResponseDto"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      security: *a24
      summary: Create an invitation
      tags: *a25
  /connections/accept-invitation:
    post:
      operationId: ConnectionController_acceptInvitation
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInvitationDto"
      responses:
        "200":
          description: Connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionRecordDto"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      security: *a24
      summary: Accept an invitation
      tags: *a25
  /connections/{id}:
    get:
      operationId: ConnectionController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionRecordDto"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security: *a24
      summary: Get a connection
      tags: *a25
...
  /credentials/{id}:
    get:
      operationId: CredentialController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Credential Record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialRecordDto"
        "401": *a26
        "404":
          description: Not Found
      security: *a27
      summary: Get a credential record
      tags: *a28
...
  /openid4vc/issuer:
    get:
      operationId: OpenId4VcIssuerController_find
      parameters:
        - name: publicIssuerId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Issuer records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenId4VcIssuerRecordDto"
        "401":
          description: Unauthorized
      security: &a31
        - bearer: []
      summary: Get issuer records
      tags: &a32
        - OpenID4VC Issuer
    post:
      operationId: OpenId4VcIssuerController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenId4VcIssuersCreateDto"
      responses:
        "200":
          description: Issuer Record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenId4VcIssuerRecordDto"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "422":
          description: Unprocessable Entity
      security: *a31
      summary: Create an Issuer
      tags: *a32
  /openid4vc/issuer/{issuerId}:
    put:
      operationId: OpenId4VcIssuerController_update
      parameters:
        - name: issuerId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenId4VcIssuersUpdateMetadataDto"
      responses:
        "200":
          description: Issuer Record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenId4VcIssuerRecordDto"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "422":
          description: Unprocessable Entity
      security: *a31
      summary: Update an Issuer Metadata
      tags: *a32
  /openid4vc/issuer/supported-credentials:
    get:
      operationId: OpenId4VcIssuerController_supportedCredentials
      parameters:
        - name: publicIssuerId
          required: true
          in: query
          schema:
            type: string
        - name: credentialType
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Supported credential
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenId4VciCredentialConfigurationSupportedWithId"
        "401":
          description: Unauthorized
      security: *a31
      summary: Get supported issuer credentials
      tags: *a32
...
info:
  title: Heka Identity Service
  description: Heka Identity Service API
  version: 2.3.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CredentialDefinitionDto:
      type: object
      properties:
        id:
          type: string
        issuerId:
          type: string
        schemaId:
          type: string
        tag:
          type: string
      required:
        - id
        - issuerId
        - schemaId
        - tag
    CreateCredentialDefinitionDto:
      type: object
      properties:
        issuerId:
          type: string
        schemaId:
          type: string
        tag:
          type: string
      required:
        - issuerId
        - schemaId
        - tag
    ConnectionRecordDto:
      type: object
      properties:
        id:
          type: string
        threadId:
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        state:
          type: string
          enum:
            - start
            - invitation-sent
            - invitation-received
            - request-sent
            - request-received
            - response-sent
            - response-received
            - abandoned
            - completed
        errorMessage:
          type: string
        role:
          type: string
          enum:
            - requester
            - responder
        did:
          type: string
        theirDid:
          type: string
        theirLabel:
          type: string
        alias:
          type: string
        invitationDid:
          type: string
      required:
        - id
        - createdAt
        - state
        - role
    CreateInvitationRequestDto:
      type: object
      properties:
        label:
          type: string
        alias:
          type: string
        imageUrl:
          type: string
        multiUseInvitation:
          type: boolean
    CreateInvitationResponseDto:
      type: object
      properties:
        id:
          type: string
        invitationUrl:
          type: string
      required:
        - id
        - invitationUrl
    AcceptInvitationDto:
      type: object
      properties:
        invitationUrl:
          type: string
        label:
          type: string
        alias:
          type: string
      required:
        - invitationUrl
    CredentialPreviewAttributeDto:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
      required:
        - name
        - value
    CredentialRecordDto:
      type: object
      properties:
        id:
          type: string
        connectionId:
          type: string
        threadId:
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        state:
          type: string
          enum:
            - proposal-sent
            - proposal-received
            - offer-sent
            - offer-received
            - declined
            - request-sent
            - request-received
            - credential-issued
            - credential-received
            - done
            - abandoned
        errorMessage:
          type: string
        credentialAttributes:
          type: array
          items:
            $ref: "#/components/schemas/CredentialPreviewAttributeDto"
      required:
        - id
        - threadId
        - createdAt
        - state
...
    OpenId4VcIssuerRecordDto:
      type: object
      properties:
        id:
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        type:
          type: string
        publicIssuerId:
          type: string
        accessTokenPublicKeyFingerprint:
          type: string
        credentialsSupported:
          type: array
          items:
            type: string
        display:
          type: array
          items:
            type: string
      required:
        - id
        - createdAt
        - type
        - publicIssuerId
        - accessTokenPublicKeyFingerprint
        - credentialsSupported
    OpenId4VciCredentialConfigurationSupportedWithId:
      type: object
      properties:
        id:
          type: string
        format:
          type: string
          enum:
            - vc+sd-jwt
            - jwt_vc_json
            - jwt_vc_json-ld
            - ldp_vc
        display:
          type: array
          items:
            type: string
        proof_types_supported:
          type: object
      required:
        - id
        - format
    OpenId4VcIssuersCreateDto:
      type: object
      properties:
        publicIssuerId:
          type: string
        credentialsSupported:
          type: array
          items:
            $ref: "#/components/schemas/OpenId4VciCredentialConfigurationSupportedWithId"
        display:
          type: array
          items:
            type: string
      required:
        - publicIssuerId
        - credentialsSupported
    OpenId4VcIssuersUpdateMetadataDto:
      type: object
      properties:
        credentialsSupported:
          items:
            type: array
          type: array
        action:
          type: string
          enum:
            - add
            - replace
        display:
          type: array
          items:
            type: string
      required:
        - credentialsSupported
...
security:
  - bearer: []